fix(droit-du-travail): move non page component to src
