+++ packages/code-du-travail-frontend/__tests__/droit-du-travail.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { render } from "@testing-library/react";
-import DroitDuTravail from "../pages/droit-du-travail/index.js";
+import DroitDuTravail from "../pages/droit-du-travail";
 
 describe("<DroitDuTravail />", () => {
   it("should render", () => {
